feat: switch to my fork of telescope-egrepify
To be clear, I didn't code anything myself, I've forked a fork and
updated it in line with the master of the official repo. This fork adds
the ability to press ctrl-h to show hidden files as desired. As such,
I've removed the "." prefix that I added previously, since it was
behaving weirdly when I searched for e.g., ".conf".