feat(cmp): don't choose completions with esc, simplify tab
Specifically, we go back to using enter, but only when a completion
option is specifically selected, otherwise we get a normal carriage
return. In this vein, we got rid of C-CR, since it was made redundant.

For tab, we simplified it so that it no longer expands snippets, it just
handles jumps. Shift-tab already worked this way, so I'm not sure why it
was the way it was honestly.